perf(json): flip lazy-tape scans to the batch parser early (#7478)
A sequential scan of a lazy JSON array was invisible to the only adaptive
signal `lazy_get` had. `cumulative_walk_steps` accumulates one step per
element on a sequential walk, against a threshold of 2n, so it provably
never trips — every element of a scanned array was materialized one at a
time at ~2.3x the batch parser's per-element rate.

Add `LazyArrayHeader::sequential_streak`, a run-length of consecutive
ascending cold reads, and trip it at `scan_flip_threshold` (n/64, floored
at 64). The flip fires while the sparse cache is still nearly empty,
which is what makes `force_materialize_lazy` choose #7499's batch reparse
rather than the element-wise merge walk — firing it late is a no-op,
which is why #7499 alone did not move `field_access`.

The trigger carries the callee's own `cached_count * 2 < cached_length`
test, so it never asks for a producer that would be declined.

Tests split into json_tape_tests.rs to stay under the 2000-line cap.
